moved validate_user into journalist_app/utils.py
