+++ securedrop/journalist.py
@@ -17,14 +17,13 @@ import i18n
 from flask_babel import gettext, ngettext
 import store
 from db import (db_session, Source, Journalist, Submission, Reply,
-                SourceStar, LoginThrottledException,
-                PasswordError, InvalidUsernameException,
-                BadTokenException, WrongPasswordException)
+                SourceStar, PasswordError, InvalidUsernameException)
 import worker
 
 from journalist_app import create_app
 from journalist_app.forms import ReplyForm
-from journalist_app.utils import logged_in, commit_account_changes, get_source
+from journalist_app.utils import (logged_in, commit_account_changes,
+                                  get_source, validate_user)
 
 app = create_app(config)
 
@@ -55,53 +54,6 @@ def admin_required(func):
               "notification")
         return redirect(url_for('index'))
     return wrapper
-
-
-def validate_user(username, password, token, error_message=None):
-    """
-    Validates the user by calling the login and handling exceptions
-    :param username: Username
-    :param password: Password
-    :param token: Two-factor authentication token
-    :param error_message: Localized error message string to use on failure
-    :return: Journalist user object if successful, None otherwise.
-    """
-    try:
-        return Journalist.login(username, password, token)
-    except (InvalidUsernameException,
-            BadTokenException,
-            WrongPasswordException,
-            LoginThrottledException) as e:
-        app.logger.error("Login for '{}' failed: {}".format(
-            username, e))
-        if not error_message:
-            error_message = gettext('Login failed.')
-        login_flashed_msg = error_message
-
-        if isinstance(e, LoginThrottledException):
-            login_flashed_msg += " "
-            period = Journalist._LOGIN_ATTEMPT_PERIOD
-            # ngettext is needed although we always have period > 1
-            # see https://github.com/freedomofpress/securedrop/issues/2422
-            login_flashed_msg += ngettext(
-                "Please wait at least {seconds} second "
-                "before logging in again.",
-                "Please wait at least {seconds} seconds "
-                "before logging in again.", period).format(seconds=period)
-        else:
-            try:
-                user = Journalist.query.filter_by(
-                    username=username).one()
-                if user.is_totp:
-                    login_flashed_msg += " "
-                    login_flashed_msg += gettext(
-                        "Please wait for a new two-factor token"
-                        " before trying again.")
-            except:
-                pass
-
-        flash(login_flashed_msg, "error")
-        return None
 
 
 @app.route('/login', methods=('GET', 'POST'))
